% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_shiny.R
\name{check_grammar_gpt}
\alias{check_grammar_gpt}
\title{Edit text for spelling, grammar, word count, clarity using GPT-3}
\usage{
check_grammar_gpt(text_to_edit, user_choice, model_type)
}
\arguments{
\item{text_to_edit:}{The original text that needs editing.}

\item{user_choice:}{A character value indicating the type of editing required. Possible values include "spelling_grammar", "word_count", or "clarity".}

\item{model_type:}{A character value indicating the GPT-3 model to use, these are supplied by the shiny app.}
}
\value{
Returns the edited text as a string.
}
\description{
This function uses GPT-3 to assist in editing text for spelling, grammar, word count, and clarity. It is a companion function for the run grammar checker Shiny app. The function takes in the following parameters:
}
