% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_shiny.R
\name{compare_text}
\alias{compare_text}
\title{Compare Two Texts}
\usage{
compare_text(original_text, modified_text)
}
\arguments{
\item{original_text}{A character string representing the original text.}

\item{modified_text}{A character string representing the modified text.}
}
\value{
Either "identical" if the two texts are identical, or a diff object showing the differences between the two texts.
}
\description{
This function compares two texts and returns whether they are identical or a diff object from diffobj to visualize the differences in HTML. Used in conjunction with the run grammar checker Shiny app.
}
