% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{completion_ada}
\alias{completion_ada}
\title{Send highlighted text as completion prompt to ada}
\usage{
completion_ada()
}
\value{
string to console
}
\description{
This is an RStudio addin function that should become available after the package is installed.
}
\details{
Highlight text in an editor window while using Rstudio, then select the addin from the drop-down menu.

This function sends the selected text to an ada model

The model receives the highlighted text, generates a completion prompt, and returns the generated text to the console.
}
