% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{continue_chat}
\alias{continue_chat}
\title{Continue a chat with message history}
\usage{
continue_chat(model_type = "gpt-4")
}
\arguments{
\item{model_type}{string, default "gpt-4"}
}
\value{
string to the editor window
}
\description{
To be used in combination with start_chat
}
\details{
Select any text in the editor then run this addin.

The selected text is combined with the existing message history and sent as a prompt to OpenAI.
}
