% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{line_editor}
\alias{line_editor}
\title{line_editor}
\usage{
line_editor()
}
\description{
This is an RStudio addin function that should become available after the package is installed.
}
\details{
Highlight text in an editor window while using Rstudio, then select the addin from the drop-down menu.

This function sends the selected text to a gpt-3.5-turbo chat model. The system prompt is:

You are an editorial writing assistant. Your task is to edit the content you receive for improved clarity, flow, and reduced word count.

The model receives the highlighted text, generates a response following the prompt, and returns to the new suggested text to the console.
}
