% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_shiny.R
\name{parse_rmd}
\alias{parse_rmd}
\title{Parse RMarkdown documents into sentences or paragraphs}
\usage{
parse_rmd(file, token_level = "sentences")
}
\arguments{
\item{file}{a character vector of the file or path}
}
\description{
Parse RMarkdown documents into sentences or paragraphs
}
\examples{
# don't run
# parse_rmd()
}
