% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{points_to_prose}
\alias{points_to_prose}
\title{Convert point form to longform prose}
\usage{
points_to_prose()
}
\value{
string to console
}
\description{
This is an RStudio addin function that should become available after the package is installed.
}
\details{
Highlight point form text in an editor window while using Rstudio, then select the addin from the drop-down menu.

This function sends the selected text to a gpt-3.5-turbo chat model. The system prompt is:

You are a writing assistant. Your task is to take point form notes and convert them into longform prose. Please write in a straightforward and clear manner. The user prompt will contain point form notes.

The model receives the highlighted text, generates a response following the prompt, and returns to the new suggested text to the console.
}
