% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{start_chat}
\alias{start_chat}
\title{Start a continuous chat with message history}
\usage{
start_chat()
}
\value{
list called my_chat to the global environment
}
\description{
Run this in the console
}
\details{
Very basic right now, no error checking. Chat history is preserved in my_chat which is returned to the global environment. Running this function initializes my_chat, and deletes anything that was in there.
}
