% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{summarize_main_points}
\alias{summarize_main_points}
\title{summarize text into a series of points}
\usage{
summarize_main_points()
}
\value{
string to console
}
\description{
This is an RStudio addin function that should become available after the package is installed.
}
\details{
Highlight text in an editor window while using Rstudio, then select the addin from the drop-down menu.

This function sends the selected text to a gpt-3.5-turbo chat model. The system prompt is:

You are a writing assistant. Your task is to summarize input text into clear pointform notes.

The model receives the highlighted text, generates a response following the prompt, and returns the new suggested text to the console.
}
