% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{write_function_doc}
\alias{write_function_doc}
\title{Generates documentation in Roxygen syntax for an R function}
\usage{
write_function_doc()
}
\value{
the documentation is printed to the console
}
\description{
This addin is registered and accessible through RStudio. The input to the function is any highighted text in the RStudio editor.
}
\examples{
#write_function_doc()

}
