% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{chat_to_doc}
\alias{chat_to_doc}
\title{Send highlighted text to gpt-4 output to doc}
\usage{
chat_to_doc(model_type = "gpt-4")
}
\arguments{
\item{model_type}{string, default "gpt-4"}
}
\value{
string to console
}
\description{
This is an RStudio addin function that should become available after the package is installed.
}
\details{
Highlight text in an editor window while using Rstudio, then select the addin from the drop-down menu.

This function sends the selected text to a gpt-4 model and returns the output to the document, not the console.

The selected text is replaced with the original text, a new line, and the response from the model.
}
