% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_shiny.R
\name{run_grammar_checker}
\alias{run_grammar_checker}
\title{Start grammar checker shiny application
Start the grammar checker.}
\usage{
run_grammar_checker(path, choose_token_level = "sentences")
}
\arguments{
\item{path}{the intended filepath}

\item{choose_token_level}{string "sentences" or "paragraphs", defaults to sentences}
}
\value{
a shiny app is launched
}
\description{
Start grammar checker shiny application
Start the grammar checker.
}
\examples{
# don't run during tests
# gptaddin::run_grammar_checker("example.rmd")
}
